+++ src/components/spartan/SInputDateBlock/SInputDateBlock.vue
@@ -1,30 +1,19 @@
 <script setup lang="ts">
-import { computed } from 'vue';
 import { SInputDate, type TInputDateProps, type TInputDateEmits } from '../SInputDate';
 import { BlockWrapper, type TBlockWrapperProps } from '@internal';
+import { extractWrapperProps } from '@/helpers';
 
 defineEmits<TInputDateEmits>();
 const props = defineProps<Partial<TBlockWrapperProps> & TInputDateProps>();
-
-const blockWrapperProps = computed(() => ({
-    label: props.label,
-    id: props.id,
-    errorText: props.errorText,
-    helpText: props.helpText,
-}));
-
-const InputDateProps = computed<TInputDateProps>(() => {
-    // @typescript-eslint/no-unused-vars
-    const { label, helpText, errorText, ...rest } = props;
-    return { ...rest, error: props.errorText ? Boolean(props.errorText) : props.error };
-});
+const [blockWrapperProps, inputDateProps] = extractWrapperProps<TInputDateProps>(props);
 </script>
 
 <template>
-    <BlockWrapper wrapper="SInputBlock" v-bind="blockWrapperProps">
+    <BlockWrapper wrapper="SInputDateBlock" v-bind="blockWrapperProps" v-slot="{ id }" useDpUid>
         <SInputDate
+            :id="id"
             class="w-full"
-            v-bind="InputDateProps"
+            v-bind="inputDateProps"
             @update:model-value="(newValue) => $emit('update:modelValue', newValue)"
         />
     </BlockWrapper>