+++ src/helpers/index.ts
@@ -3,3 +3,4 @@ export * from './i18n';
 export * from './slotContent';
 export * from './styles';
 export * from './sanitizer';
+export * from './blockWrapper';